requirements.txt
alembic==1.7.3
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
PyJWT==1.4.2
PyMySQL==1.0.2
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
Werkzeug==0.16.1
zipp==3.5.0
